+++ stwo_cairo_prover/crates/prover/src/adapter/vm_import/mod.rs
@@ -14,7 +14,7 @@ use tracing::{span, Level};
 
 use super::builtins::BuiltinSegments;
 use super::memory::MemoryConfig;
-use super::state_transitions::StateTransitions;
+use super::opcodes::StateTransitions;
 use super::ProverInput;
 use crate::adapter::memory::MemoryBuilder;
 use crate::components::memory::MEMORY_ADDRESS_BOUND;
