+++ stwo_cairo_prover/crates/prover/src/cairo_air/opcodes_air.rs
@@ -12,7 +12,7 @@ use stwo_prover::core::pcs::{TreeBuilder, TreeVec};
 
 use super::air::CairoInteractionElements;
 use super::debug_tools::display_components;
-use crate::adapter::state_transitions::StateTransitions;
+use crate::adapter::opcodes::StateTransitions;
 use crate::components::{
     add_ap_opcode, add_ap_opcode_imm, add_ap_opcode_op_1_base_fp, add_opcode, add_opcode_imm,
     add_opcode_small, add_opcode_small_imm, assert_eq_opcode, assert_eq_opcode_double_deref,
